Move _KinematicsBuilder to new location
